Update README.md
    Overview

In this project, you will implement a multiprocessor operating system simulator using a popular threading library for Linux called pthreads. The framework for the multiprocessor OS simulator is nearly complete, but missing one critical component: the process scheduler! Your task is to implement the process scheduler and three different scheduling algorithms.

The simulated operating system supports only one thread per process making it similar to the systems that we discussed in Chapter 6. However, the simulator itself will use a thread to represent each of the CPUs in the simulated hardware. This means that the CPUs in the simulator will appear to operate concurrently.

Note: Multiple CPU cores need to be enabled for this to work correctly. The CS 2200 Docker container should already be configured to run with 4 cores. Please see a TA if you are running into any problems.

We have provided you with source files that constitute the framework for your simulator. You will only need to modify answers.txt and student.c. However, just because you are only modifying two files doesn’t mean that you should ignore the other ones.

We have provided you the following files:

    os-sim.c – Code for the operating system simulator which calls your CPU scheduler.

    os-sim.h – Header file for the simulator.

    process.c – Descriptions of the simulated processes.

    process.h – Header file for the process data.

    student.c – This file contains stub functions for your CPU scheduler.

    student.h – Header file for your code to interface with the OS simulator. Also contains the ready queue struct definition.

    answers.txt – This is a text file that you should use to write your answers to the questions listed throughout the PDF.